Group viewers into a single module.
