+++ src/gpsea/view/__init__.py
@@ -1,23 +1,28 @@
-from ._cohort import CohortViewer
-from ._cohort_variant_viewer import CohortVariantViewer
-from ._disease import DiseaseViewer
 from ._phenotype_analysis import summarize_hpo_analysis
-from ._protein_viewer import ProteinVariantViewer
 from ._protein_visualizable import ProteinVisualizable
 from ._base import GpseaReport
-from ._stats import MtcStatsViewer
 from ._txp import VariantTranscriptVisualizer
+from ._viewers import (
+    CohortVariantViewer,
+    CohortViewer,
+    DiseaseViewer,
+    MtcStatsViewer,
+    ProteinVariantViewer,
+)
 from ._protein_visualizer import ProteinVisualizer
 from ._formatter import Formatter, VariantFormatter
 
 __all__ = [
-    'GpseaReport',
-    'CohortVariantViewer',
-    'CohortViewer',
-    'ProteinVisualizer', 'ProteinVisualizable', 'ProteinVariantViewer',
-    'DiseaseViewer',
-    'MtcStatsViewer',
-    'summarize_hpo_analysis',
-    'VariantTranscriptVisualizer',
-    'Formatter', 'VariantFormatter',
+    "GpseaReport",
+    "CohortVariantViewer",
+    "CohortViewer",
+    "ProteinVisualizer",
+    "ProteinVisualizable",
+    "ProteinVariantViewer",
+    "DiseaseViewer",
+    "MtcStatsViewer",
+    "summarize_hpo_analysis",
+    "VariantTranscriptVisualizer",
+    "Formatter",
+    "VariantFormatter",
 ]
