--- tests/view/test_variant_viewer.py
@@ -1,16 +1,0 @@
-import pytest
-
-from gpsea.model import Cohort
-from gpsea.view import CohortVariantViewer
-
-
-@pytest.mark.skip("For manual run only")
-def test_viewer(
-    suox_mane_tx_id: str,
-    suox_cohort: Cohort,
-):
-    viewer = CohortVariantViewer(tx_id=suox_mane_tx_id)
-    html = viewer.process(suox_cohort)
-
-    with open("all_variants.html", "w") as fh:
-        fh.write(html)
